+++ src/Endpoints/Clusters.php
@@ -4,9 +4,11 @@ namespace Vdhicts\Cyberfusion\ClusterApi\Endpoints;
 
 use Vdhicts\Cyberfusion\ClusterApi\Exceptions\RequestException;
 use Vdhicts\Cyberfusion\ClusterApi\Models\Cluster;
+use Vdhicts\Cyberfusion\ClusterApi\Models\TaskCollection;
 use Vdhicts\Cyberfusion\ClusterApi\Request;
 use Vdhicts\Cyberfusion\ClusterApi\Response;
 use Vdhicts\Cyberfusion\ClusterApi\Support\ListFilter;
+use Vdhicts\Cyberfusion\ClusterApi\Support\Str;
 
 class Clusters extends Endpoint
 {
@@ -70,11 +72,16 @@ class Clusters extends Endpoint
      * @return Response
      * @throws RequestException
      */
-    public function commit(int $id): Response
+    public function commit(int $id, string $callbackUrl = null): Response
     {
+        $url = Str::optionalQueryParameters(
+            'cluster-deployments',
+            ['callback_url' => $callbackUrl]
+        );
+
         $request = (new Request())
             ->setMethod(Request::METHOD_POST)
-            ->setUrl('cluster-deployments')
+            ->setUrl($url)
             ->setBody([
                 'cluster_id' => $id,
             ]);
@@ -87,7 +94,7 @@ class Clusters extends Endpoint
         }
 
         return $response->setData([
-            'cluster' => (new Cluster())->fromArray($response->getData()),
+            'taskCollection' => (new TaskCollection())->fromArray($response->getData()),
         ]);
     }
 }