+++ CHANGELOG.md
@@ -7,6 +7,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 cluster API. See the changelog of the [cluster API](https://cluster-api.cyberfusion.nl/redoc#section/Changelog) for 
 detailed information.
 
+## [1.48.0]
+
+### Added
+
+- Add callback url for deployments.
+- Add callback url for the automatic deployments.
+
+### Changed
+
+- Update to [API version 1.117](https://cluster-api.cyberfusion.nl/redoc#section/Changelog/1.117-2022-03-10).
+- A commit will now return a task collection.
+
 ## [1.47.1]
 
 ### Fixed